#30 squares are clickable
